Do not manage customized auth.conf file
Prior to this commit, the auth.conf file would be managed even if the
user had manually added custom rules in the file.  This commit
introduces a fact, custom_auth_conf, that is true if the user has
manually modified the auth.conf. If this fact is true, a warning is
printed for each auth_conf::acl resource being managed.