Update manifest to use is_pe fact and str2bool
Previously the init.pp would munge the puppetversion fact to determine whether
the node was a PE node or not. As of version 2.5.1, stdlib has an is_pe fact
that will accomplish the same thing. This commit updates the manifest to use
the $is_pe fact after being turned into a boolean via str2bool. It also updates
the modulefile with a stdlib dependency of >= 2.5.1 to ensure the is_pe fact is
available.