Make sure inventory is a boolean before using it
Previously the $real_inventory was set using a selector with 'true' and 'false'
cases for $inventory, but $inventory defaulted to false, which wasn't in the
selector list. This would cause the module to fail on split master/console
installs. This updates the setting of $real_inventory to use str2bool if
$inventory is a string, otherwise match the boolean value with a boolean value,
otherwise default to false.