Properly detect modified auth_conf
Previous to this commit, the custom_auth_conf facter fact did not
properly detect a modified auth.conf file.  First, the conditional
always returned false if the auth.conf file existed.  Second, if the
SHA1 of a stripped auth.conf file matched a known SHA1, the conditional
returned true when it should have returned false.  Similarly, it
returned false when it should have returned true.  This commit fixes the
conditionals