(maint) update custom fact to strip comments
Prior to this commit, the custom fact would take into account the
comments when determining if an auth.conf was modified. This commit adds
logic to strip those comments out from what is considered a change. As a
result, the SHA1s of the unmodified files required an update. This
commit also removes the logic and test for the 'Managed by Puppet'
header.