Fix bug:78. Add split master 2.7.0 default SHA
Prior to this commit, the custom fact reported that auth.conf had been
modified for a split master/console install on 2.7.0. This incorrect
behaviour is the result of the inventory stanzas (/fact) existing in the
file used to generate the SHA for the master's auth.conf. This commit
retains that SHA but notes that it's actually a default *console* and
also adds the correct SHA for the master's auth.conf.