(PE-581) Remove real_inventory logic from auth_conf module
In PE 3, we are removing the fake master that is up only to serve inventory
requests. This means that only the master needs an auth.conf, and the
save-facts path is not needed, because the master is doing the saving. The find
and search methods on fact have been moved to the request_manager module, as
they are pe specific.